[FIX] Changed names of Protein-Protein XL-MS tools
Changed names of the Protein-Protein cross-linking tool executables to reflect the official name change of the project
