+++ src/openms/source/ANALYSIS/XLMS/sources.cmake
@@ -3,7 +3,7 @@ set(directory source/ANALYSIS/XLMS)
 
 ### list all filenames of the directory here
 set(sources_list
-OpenXQuestScores.cpp
+OpenProXLUtils.cpp
 )
 
 ### add path to the filenames
